This is the first script I am submitting.
I have used this script personally for a while now. Because of the new rules, I must now submit it to the whitelist or face being banned from league play. I also sometimes change the message that is displayed, the color of the text, and the location of the text, so will I also need to submit each variation?